\name{bookProperties-package}
\alias{bookProperties-package}
\alias{bookProperties}
\docType{package}
\title{
\packageTitle{bookProperties}
}
\description{
\packageDescription{bookProperties}
}
\author{
\packageAuthor{bookProperties}

Maintainer: \packageMaintainer{bookProperties}
}

\keyword{NLP}


