% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCharacters.R
\name{extractCharacters}
\alias{extractCharacters}
\title{extractCharacters}
\usage{
extractCharacters(book)
}
\arguments{
\item{book}{Stagger tagged book}
}
\value{
Returns extracted characters
}
\description{
A function that extracts characters from a text
}
