% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeelings.R
\name{extractFeelings}
\alias{extractFeelings}
\title{extractFeelings}
\usage{
extractFeelings(text, path_sentiment_lexicon, language)
}
\arguments{
\item{text}{Character vector with text}

\item{path_sentiment_lexicon}{Path to folder where sentiment lexicon is stored}

\item{language}{Book language}
}
\value{
Returns the predicted feelings of the text
}
\description{
A function that predicts the feelings of a text
}
