% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLocations.R
\name{extractLocations}
\alias{extractLocations}
\title{extractLocations}
\usage{
extractLocations(book, path_geo, language)
}
\arguments{
\item{book}{Stagger tagged book}

\item{path_geo}{Path to folder where geographical data is stored}

\item{language}{Book language}
}
\value{
Returns a a vector with locations and corresponding countries/continents as well as their frequencies
}
\description{
A function that extract locations and corresponding countries/continents from a book
}
