% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractReadabilityInformation.R
\name{extractReadabilityInformation}
\alias{extractReadabilityInformation}
\title{extractReadabilityInformation}
\usage{
extractReadabilityInformation(book, genre, path_models, path_genre_prop,
  readability, language)
}
\arguments{
\item{book}{Stagger tagged book}

\item{path_models}{Path to folder where glm models are stored}

\item{path_genre_prop}{Path to folder where genre probabilities are stored}

\item{readability}{Vector containing N_words, N_unique_words, N_long_words, N_sentences, LIX}

\item{language}{Book language}
}
\value{
Returns a vector with readability information, such as number of words and LIX
}
\description{
A function that extracts readability information from a book
}
