% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestEmotionsBasedOnGenre.R
\name{suggestFeelingsBasedOnGenre}
\alias{suggestFeelingsBasedOnGenre}
\title{suggestEmotionsBasedOnGenre}
\usage{
suggestFeelingsBasedOnGenre(genre, path_emotions_genre_prop, language)
}
\arguments{
\item{genre}{Book genre}

\item{path_emotions_genre_prop}{Path to folder where emotion probabilities are stored}

\item{language}{Book language}
}
\value{
Returns the predicted feelings of the text
}
\description{
A function that suggests the emotions of a text based on book genre
}
