% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestSubGenres.R
\name{suggestSubGenres}
\alias{suggestSubGenres}
\title{suggestSubGenres}
\usage{
suggestSubGenres(genre, path_sub_genre_prob, language)
}
\arguments{
\item{genre}{Book genre}

\item{language}{Book language}

\item{path_sub_genre_prop}{Path to folder where sub genre probabilities are stored}
}
\value{
Returns suggested sub genres
}
\description{
A function that suggests sub genres based on book genre
}
