% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggestSynonymEmotions.R
\name{suggestSynonymEmotions}
\alias{suggestSynonymEmotions}
\title{suggestSynonymEmotions}
\usage{
suggestSynonymEmotions(emotion_word, path_emotion_synonyms, language)
}
\arguments{
\item{emotion_word}{Emotion word}

\item{path_emotion_synonyms}{Path to folder where emotion synonyms are stored}

\item{language}{Book language}
}
\value{
Returns synonym emotions
}
\description{
A function that suggests synonym emotions of an emotion word
}
