\name{targetAge-package}
\alias{targetAge-package}
\alias{targetAge}
\docType{package}
\title{
\packageTitle{targetAge}
}
\description{
\packageDescription{targetAge}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{targetAge}
\packageIndices{targetAge}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{targetAge}

Maintainer: \packageMaintainer{targetAge}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
