\name{bookProperties-package}
\alias{bookProperties-package}
\alias{bookProperties}
\docType{package}
\title{
\packageTitle{bookProperties}
}
\description{
\packageDescription{bookProperties}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bookProperties}
\packageIndices{bookProperties}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bookProperties}

Maintainer: \packageMaintainer{bookProperties}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
